+++ client/web/src/components/Timeline.tsx
@@ -1,6 +1,6 @@
 import React, { FunctionComponent, ReactNode, useState } from 'react'
 
-import { mdiChevronDown, mdiChevronRight } from '@mdi/js'
+import { mdiChevronDown, mdiChevronUp } from '@mdi/js'
 import VisuallyHidden from '@reach/visually-hidden'
 import classNames from 'classnames'
 import { formatDistance } from 'date-fns/esm'
@@ -84,7 +84,7 @@ const TimelineStage: FunctionComponent<React.PropsWithChildren<TimelineStageProp
                 className="p-0 m-0 border-0 w-100 font-weight-normal d-flex justify-content-between align-items-center"
             >
                 {stageLabel}
-                <Icon aria-hidden={true} svgPath={isExpanded ? mdiChevronDown : mdiChevronRight} className="mr-1" />
+                <Icon aria-hidden={true} svgPath={isExpanded ? mdiChevronUp : mdiChevronDown} className="mr-1" />
             </CollapseHeader>
             <CollapsePanel className={styles.details}>{details}</CollapsePanel>
         </Collapse>