+++ client/web/src/enterprise/batches/create/ConfigurationForm.tsx
@@ -5,11 +5,10 @@ import classNames from 'classnames'
 import { noop } from 'lodash'
 import { useHistory, useLocation } from 'react-router'
 
-import { ErrorAlert } from '@sourcegraph/branded/src/components/alerts'
 import { Form } from '@sourcegraph/branded/src/components/Form'
 import { useMutation } from '@sourcegraph/http-client'
 import { SettingsOrgSubject, SettingsUserSubject } from '@sourcegraph/shared/src/settings/settings'
-import { Alert, Button, Container, Icon, Input, RadioButton, Tooltip } from '@sourcegraph/wildcard'
+import { Alert, Button, Container, Icon, Input, RadioButton, Tooltip, ErrorAlert } from '@sourcegraph/wildcard'
 
 import { AuthenticatedUser } from '../../../auth'
 import {