+++ client/web/src/enterprise/insights/pages/dashboards/dashboard-view/components/add-insight-modal/components/add-insight-modal-content/AddInsightModalContent.tsx
@@ -2,8 +2,7 @@ import React from 'react'
 
 import { escapeRegExp } from 'lodash'
 
-import { ErrorAlert } from '@sourcegraph/branded/src/components/alerts'
-import { Button, Input, Link, Label, Checkbox } from '@sourcegraph/wildcard'
+import { Button, Input, Link, Label, Checkbox, ErrorAlert } from '@sourcegraph/wildcard'
 
 import { LoaderButton } from '../../../../../../../../../components/LoaderButton'
 import { AccessibleInsight } from '../../../../../../../../../graphql-operations'