+++ cmd/frontend/graphqlbackend/insights.go
@@ -22,6 +22,9 @@ type InsightsResolver interface {
 	SearchInsightLivePreview(ctx context.Context, args SearchInsightLivePreviewArgs) ([]SearchInsightLivePreviewSeriesResolver, error)
 	SearchInsightPreview(ctx context.Context, args SearchInsightPreviewArgs) ([]SearchInsightLivePreviewSeriesResolver, error)
 
+	ValidateScopedInsightQuery(ctx context.Context, args ValidateScopedInsightQueryArgs) (ScopedInsightQueryPayloadResolver, error)
+	PreviewRepositoriesFromQuery(ctx context.Context, args PreviewRepositoriesFromQueryArgs) (RepositoryPreviewPayloadResolver, error)
+
 	// Mutations
 	CreateInsightsDashboard(ctx context.Context, args *CreateInsightsDashboardArgs) (InsightsDashboardPayloadResolver, error)
 	UpdateInsightsDashboard(ctx context.Context, args *UpdateInsightsDashboardArgs) (InsightsDashboardPayloadResolver, error)
@@ -201,6 +204,8 @@ type InsightViewResolver interface {
 	AppliedSeriesDisplayOptions(ctx context.Context) (InsightViewSeriesDisplayOptionsResolver, error)
 	Dashboards(ctx context.Context, args *InsightsDashboardsArgs) InsightsDashboardConnectionResolver
 	SeriesCount(ctx context.Context) (*int32, error)
+	RepositoryDefinition(ctx context.Context) (InsightRepositoryDefinition, error)
+	TimeScope(ctx context.Context) (InsightTimeScope, error)
 }
 
 type InsightDataSeriesDefinition interface {
@@ -227,6 +232,7 @@ type SearchInsightDataSeriesDefinitionResolver interface {
 	SeriesId(ctx context.Context) (string, error)
 	Query(ctx context.Context) (string, error)
 	RepositoryScope(ctx context.Context) (InsightRepositoryScopeResolver, error)
+	RepositoryDefinition(ctx context.Context) (InsightRepositoryDefinition, error)
 	TimeScope(ctx context.Context) (InsightTimeScope, error)
 	GeneratedFromCaptureGroups() (bool, error)
 	IsCalculated() (bool, error)
@@ -249,6 +255,16 @@ type InsightIntervalTimeScope interface {
 
 type InsightRepositoryScopeResolver interface {
 	Repositories(ctx context.Context) ([]string, error)
+}
+
+type InsightRepositoryDefinition interface {
+	ToInsightRepositoryScope() (InsightRepositoryScopeResolver, bool)
+	ToRepositorySearchScope() (RepositorySearchScopeResolver, bool)
+}
+
+type RepositorySearchScopeResolver interface {
+	Search() string
+	AllRepositories() bool
 }
 
 type InsightsDashboardPayloadResolver interface {
@@ -322,10 +338,12 @@ type CreateLineChartSearchInsightArgs struct {
 }
 
 type CreateLineChartSearchInsightInput struct {
-	DataSeries   []LineChartSearchInsightDataSeriesInput
-	Options      LineChartOptionsInput
-	Dashboards   *[]graphql.ID
-	ViewControls *InsightViewControlsInput
+	DataSeries      []LineChartSearchInsightDataSeriesInput
+	Options         LineChartOptionsInput
+	Dashboards      *[]graphql.ID
+	ViewControls    *InsightViewControlsInput
+	RepositoryScope *RepositoryScopeInput
+	TimeScope       *TimeScopeInput
 }
 
 type UpdateLineChartSearchInsightArgs struct {
@@ -337,6 +355,8 @@ type UpdateLineChartSearchInsightInput struct {
 	DataSeries          []LineChartSearchInsightDataSeriesInput
 	PresentationOptions LineChartOptionsInput
 	ViewControls        InsightViewControlsInput
+	RepositoryScope     *RepositoryScopeInput
+	TimeScope           *TimeScopeInput
 }
 
 type CreatePieChartSearchInsightArgs struct {
@@ -400,8 +420,8 @@ type InsightViewFiltersInput struct {
 type LineChartSearchInsightDataSeriesInput struct {
 	SeriesId                   *string
 	Query                      string
-	TimeScope                  TimeScopeInput
-	RepositoryScope            RepositoryScopeInput
+	TimeScope                  *TimeScopeInput
+	RepositoryScope            *RepositoryScopeInput
 	Options                    LineChartDataSeriesOptionsInput
 	GeneratedFromCaptureGroups *bool
 	GroupBy                    *string
@@ -413,7 +433,8 @@ type LineChartDataSeriesOptionsInput struct {
 }
 
 type RepositoryScopeInput struct {
-	Repositories []string
+	Repositories       []string
+	RepositoryCriteria *string
 }
 
 type TimeScopeInput struct {
@@ -475,4 +496,23 @@ type TimeoutDatapointAlert interface {
 type GenericIncompleteDatapointAlert interface {
 	Time() gqlutil.DateTime
 	Reason() string
+}
+
+type ValidateScopedInsightQueryArgs struct {
+	Query string
+}
+
+type ScopedInsightQueryPayloadResolver interface {
+	Query(ctx context.Context) string
+	IsValid(ctx context.Context) bool
+	InvalidReason(ctx context.Context) *string
+}
+
+type PreviewRepositoriesFromQueryArgs struct {
+	Query string
+}
+
+type RepositoryPreviewPayloadResolver interface {
+	Query(ctx context.Context) string
+	NumberOfRepositories(ctx context.Context) *int32
 }