+++ dev/scaletesting/codehostcopy/runner.go
@@ -68,7 +68,51 @@ func (r *Runner) addSSHKey(ctx context.Context) (func(), error) {
 	}, nil
 }
 
-func (r *Runner) Run(ctx context.Context, concurrency int) error {
+func (r *Runner) List(ctx context.Context, limit int) error {
+	out := output.NewOutput(os.Stdout, output.OutputOpts{})
+
+	// Load existing repositories.
+	srcRepos, err := r.store.Load()
+	if err != nil {
+		r.logger.Error("failed to open state database", log.Error(err))
+		return err
+	}
+	loadedFromDB := true
+
+	// If we're starting fresh, really fetch them.
+	if len(srcRepos) == 0 {
+		loadedFromDB = false
+		r.logger.Info("No existing state found, creating ...")
+		out.WriteLine(output.Line(output.EmojiHourglass, output.StyleBold, "Listing repos"))
+
+		var repos []*store.Repo
+		repoIter := r.source.Iterator()
+		for !repoIter.Done() && repoIter.Err() == nil {
+			repos = append(repos, repoIter.Next(ctx)...)
+		}
+
+		if err != nil {
+			r.logger.Error("failed to list repositories from source", log.Error(err))
+			return err
+		}
+		srcRepos = repos
+		if err := r.store.Insert(repos); err != nil {
+			r.logger.Error("failed to insert repositories from source", log.Error(err))
+			return err
+		}
+	}
+	block := out.Block(output.Line(output.EmojiInfo, output.StyleBold, fmt.Sprintf("List of repos (db: %v limit: %d total: %d)", loadedFromDB, limit, len(srcRepos))))
+	if limit != 0 && limit < len(srcRepos) {
+		srcRepos = srcRepos[:limit]
+	}
+	for _, r := range srcRepos {
+		block.Writef("Name: %s Created: %v Pushed: %v GitURL: %s ToGitURL: %s Failed: %s", r.Name, r.Created, r.Pushed, r.GitURL, r.ToGitURL, r.Failed)
+	}
+	block.Close()
+	return nil
+}
+
+func (r *Runner) Copy(ctx context.Context, concurrency int) error {
 	out := output.NewOutput(os.Stdout, output.OutputOpts{})
 
 	out.WriteLine(output.Line(output.EmojiInfo, output.StyleGrey, "Adding codehost ssh key"))
@@ -122,37 +166,39 @@ func (r *Runner) Run(ctx context.Context, concurrency int) error {
 		}
 
 		for _, repo := range repos {
-			currentRepo := repo
+			repo := repo
 			g.Go(func() error {
 				// Create the repo on destination.
-				if !currentRepo.Created {
-					toGitURL, cErr := r.destination.CreateRepo(ctx, currentRepo.Name)
-					if cErr != nil {
-						currentRepo.Failed = cErr.Error()
-						r.logger.Error("failed to create repo", logRepo(currentRepo, log.Error(cErr))...)
+				if !repo.Created {
+					toGitURL, err := r.destination.CreateRepo(ctx, repo.Name)
+					if err != nil {
+						repo.Failed = err.Error()
+						r.logger.Error("failed to create repo", logRepo(repo, log.Error(err))...)
 					} else {
-						currentRepo.ToGitURL = toGitURL.String()
-						currentRepo.Created = true
+						repo.ToGitURL = toGitURL.String()
+						repo.Created = true
+						// If we resumed and this repo previously failed, we need to clear the failed status as it succeeded now
+						repo.Failed = ""
 					}
-					if cErr = r.store.SaveRepo(currentRepo); cErr != nil {
-						r.logger.Error("failed to save repo", logRepo(currentRepo, log.Error(cErr))...)
-						return cErr
+					if err = r.store.SaveRepo(repo); err != nil {
+						r.logger.Error("failed to save repo", logRepo(repo, log.Error(err))...)
+						return err
 					}
 				}
 
 				// Push the repo on destination.
-				if !currentRepo.Pushed && currentRepo.Created {
-					cErr := pushRepo(ctx, currentRepo, r.source.GitOpts(), r.destination.GitOpts())
+				if !repo.Pushed && repo.Created {
+					err := pushRepo(ctx, repo, r.source.GitOpts(), r.destination.GitOpts())
 					// state might be out of date so ignore existing repos
-					if cErr != nil && !strings.Contains(cErr.Error(), "has already been taken") {
-						currentRepo.Failed = cErr.Error()
-						r.logger.Error("failed to push repo", logRepo(currentRepo, log.Error(cErr))...)
+					if err != nil && !strings.Contains(err.Error(), "has already been taken") {
+						repo.Failed = err.Error()
+						r.logger.Error("failed to push repo", logRepo(repo, log.Error(err))...)
 					} else {
-						currentRepo.Pushed = true
+						repo.Pushed = true
 					}
-					if cErr = r.store.SaveRepo(currentRepo); cErr != nil {
-						r.logger.Error("failed to save repo", logRepo(currentRepo, log.Error(cErr))...)
-						return cErr
+					if err = r.store.SaveRepo(repo); err != nil {
+						r.logger.Error("failed to save repo", logRepo(repo, log.Error(err))...)
+						return err
 					}
 				}
 				atomic.AddInt64(&done, 1)
@@ -162,6 +208,11 @@ func (r *Runner) Run(ctx context.Context, concurrency int) error {
 			})
 		}
 	}
+
+	if repoIter.Err() != nil {
+		return repoIter.Err()
+	}
+
 	errs := g.Wait()
 	for _, e := range errs {
 		if e != nil {
@@ -172,6 +223,11 @@ func (r *Runner) Run(ctx context.Context, concurrency int) error {
 }
 
 func pushRepo(ctx context.Context, repo *store.Repo, srcOpts []GitOpt, destOpts []GitOpt) error {
+	// Handle the testing case
+	if strings.HasPrefix(repo.ToGitURL, "https://dummy.local") {
+		return nil
+	}
+
 	tmpDir, err := os.MkdirTemp(os.TempDir(), fmt.Sprintf("repo__%s", repo.Name))
 	if err != nil {
 		return err
@@ -180,7 +236,9 @@ func pushRepo(ctx context.Context, repo *store.Repo, srcOpts []GitOpt, destOpts 
 		_ = os.RemoveAll(tmpDir)
 	}()
 
-	cmd := run.Bash(ctx, "git clone", repo.GitURL).Dir(tmpDir)
+	// we add the repo name so that we ensure we cd to the right repo directory
+	// if we don't do this, there is no guarantee that the repo name and the git url are the same
+	cmd := run.Bash(ctx, "git clone --mirror", repo.GitURL).Dir(tmpDir)
 	for _, opt := range srcOpts {
 		cmd = opt(cmd)
 	}
@@ -203,7 +261,8 @@ func pushRepo(ctx context.Context, repo *store.Repo, srcOpts []GitOpt, destOpts 
 func gitPushWithRetry(ctx context.Context, dir string, retry int, destOpts ...GitOpt) error {
 	var err error
 	for i := 0; i < retry; i++ {
-		cmd := run.Bash(ctx, "git push destination").Dir(dir)
+		// --force, with mirror we want the remote to look exactly as we have it
+		cmd := run.Bash(ctx, "git push --mirror --force destination").Dir(dir)
 		for _, opt := range destOpts {
 			cmd = opt(cmd)
 		}