+++ enterprise/internal/codeintel/autoindexing/internal/inference/lua/go.lua
@@ -6,6 +6,15 @@ local shared = require "sg.autoindex.shared"
 
 local indexer = require("sg.autoindex.indexes").get "go"
 
+local netrc_steps = [[
+if [ "$NETRC_DATA" ]; then
+  echo "Writing netrc config to $HOME/.netrc"
+  echo "$NETRC_DATA" > ~/.netrc
+else
+  echo "No netrc config set, continuing"
+fi
+]]
+
 local exclude_paths = pattern.new_path_combine(shared.exclude_paths, {
   pattern.new_path_segment "vendor",
 })
@@ -27,13 +36,15 @@ local gomod_recognizer = recognizer.new_path_recognizer {
           {
             root = root,
             image = indexer,
-            commands = { "go mod download" },
+            commands = { netrc_steps, "go mod download" },
           },
         },
+        local_steps = { netrc_steps },
         root = root,
         indexer = indexer,
         indexer_args = { "lsif-go", "--no-animation" },
         outfile = "",
+        requested_envvars = { "GOPRIVATE", "GOPROXY", "GONOPROXY", "GOSUMDB", "GONOSUMDB", "NETRC_DATA" },
       })
     end
 
@@ -54,11 +65,12 @@ local goext_recognizer = recognizer.new_path_recognizer {
     for i = 1, #paths do
       if path.dirname(paths[i]) == "" then
         return {
-          steps = {},
+          local_steps = { netrc_steps },
           root = "",
           indexer = indexer,
           indexer_args = { "GO111MODULE=off", "lsif-go", "--no-animation" },
           outfile = "",
+          requested_envvars = { "GOPRIVATE", "GOPROXY", "GONOPROXY", "GOSUMDB", "GONOSUMDB", "NETRC_DATA" },
         }
       end
     end