+++ internal/database/database.go
@@ -36,6 +36,7 @@ type DB interface {
 	OrgMembers() OrgMemberStore
 	Orgs() OrgStore
 	OrgStats() OrgStatsStore
+	Permissions() PermissionStore
 	Phabricator() PhabricatorStore
 	Repos() RepoStore
 	RepoKVPs() RepoKVPStore
@@ -48,6 +49,7 @@ type DB interface {
 	UserCredentials(encryption.Key) UserCredentialsStore
 	UserEmails() UserEmailsStore
 	UserExternalAccounts() UserExternalAccountsStore
+	UserRoles() UserRoleStore
 	Users() UserStore
 	WebhookLogs(encryption.Key) WebhookLogStore
 	Webhooks(encryption.Key) WebhookStore
@@ -166,6 +168,10 @@ func (d *db) OrgStats() OrgStatsStore {
 	return OrgStatsWith(d.Store)
 }
 
+func (d *db) Permissions() PermissionStore {
+	return &permissionStore{Store: d.Store}
+}
+
 func (d *db) Phabricator() PhabricatorStore {
 	return PhabricatorWith(d.Store)
 }
@@ -179,7 +185,7 @@ func (d *db) RepoKVPs() RepoKVPStore {
 }
 
 func (d *db) Roles() RoleStore {
-	return &roleStore{d.Store}
+	return RolesWith(d.Store)
 }
 
 func (d *db) SavedSearches() SavedSearchStore {
@@ -212,6 +218,10 @@ func (d *db) UserEmails() UserEmailsStore {
 
 func (d *db) UserExternalAccounts() UserExternalAccountsStore {
 	return ExternalAccountsWith(d.logger, d.Store)
+}
+
+func (d *db) UserRoles() UserRoleStore {
+	return UserRolesWith(d.Store)
 }
 
 func (d *db) Users() UserStore {